+++ survey-parent/survey-user-impl/src/main/java/com/lego/survey/user/impl/service/impl/LogServiceImpl.java
@@ -56,7 +56,7 @@ public class LogServiceImpl implements ILogService {
             }
             logVos.add(logVo);
         }
-        pagedResult.setPage(new com.survey.lib.common.page.Page(pageIndex, pageSize, 0, logPage.getTotalElements(), logPage.getTotalPages()));
+        pagedResult.setPage(new com.survey.lib.common.page.Page(pageIndex, pageSize, 0, (int)logPage.getTotalElements(), logPage.getTotalPages()));
         pagedResult.setResultList(logVos);
         return RespVOBuilder.success(pagedResult);
     }
