+++ survey-parent/survey-user-impl/src/main/java/com/lego/survey/user/impl/service/impl/UserServiceImpl.java
@@ -330,7 +330,7 @@ public class UserServiceImpl implements IUserService {
         Long count = mongoTemplate.count(query, Long.class, tableName);
         pagedResult.setResultList(objectList);
         int totalPage = count.intValue() % pageSize == 0 ? count.intValue() / pageSize : count.intValue() / pageSize + 1;
-        pagedResult.setPage(new com.survey.lib.common.page.Page(pageIndex, pageSize, 0, count, totalPage));
+        pagedResult.setPage(new com.survey.lib.common.page.Page(pageIndex, pageSize, 0,count.intValue(), totalPage));
         return pagedResult;
     }
